Update README.md
As an undergraduate from Open University of Sri Lanka, Nawala, in the stream of IT, I look for an opportunity to exposure me to the industry and to practice the theoretical studies in the field of Information Technology. I am very much enthusiastic, self-motivated, hardworking team player and committed to support the goals and aims of the company.